fix(pwm): cast pwm_capture_timer_index's identical 0/1 narrowing
Same -Wconversion shape as pwm_timer_index() in pwm.c, in the sibling
capture-mirror index helper -- not covered by #109's pwm.c-only
override, and this file is compiled by the gd32 backend build for the
first time in this change. Same fix: explicit cast, values 0/1 always
fit uint8_t.

+++ hal/gd32/pwm_capture.c
@@ -132,7 +132,11 @@ static bool     pwm_capture_pending_valid[2];
 /* TIMER base -> pwm_capture_active_car / pending index. */
 static uint8_t pwm_capture_timer_index(uint32_t periph)
 {
-	return (periph == TIMER0) ? 0u : 1u;
+	/* Same shape as pwm.c's pwm_timer_index(): the ternary promotes to
+	 * unsigned int before the uint8_t return narrows it, and -Wconversion
+	 * can't see that only 0 or 1 ever comes out. Both fit; explicit cast
+	 * only to satisfy the warning. */
+	return (uint8_t)((periph == TIMER0) ? 0u : 1u);
 }
 
 /* Consume ONE outstanding update event for periph's timer, if
